requirements.txt
python-telegram-bot[job-queue, rate-limiter]==20.1
PyYAML==6.0
pymongo==4.3.3
python-dotenv==0.21.0
yt-dlp